backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas>=2.2.3
numpy>=2.1.3
python-dotenv==1.0.1
requests==2.32.3
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
motor>=3.4.0
pymongo>=4.7.0
